requirements.txt
google-search-results==2.4.2

# Web Scraping
beautifulsoup4>=4.9.3
selenium>=4.0.0
requests>=2.26.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.1

# API Integration
openai>=0.27.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0

# Web Framework & Frontend
flask>=2.0.0
streamlit>=1.0.0
jinja2>=3.0.0

# Email
python-dotenv>=0.19.0
email-validator>=1.1.0

# Utils
pandas>=1.3.0
tqdm>=4.62.0
loguru>=0.5.0

# Optional - Translation
deepl==1.16.0
googletrans>=3.0.0

# Testing
pytest>=6.2.0

# Development
black>=21.5b2
flake8>=3.9.0
isort>=5.9.0
undetected-chromedriver
instagram_private_api
